intel_sgx: driver for Intel Secure Guard eXtensions
Intel(R) SGX is a set of CPU instructions that can be used by
applications to set aside private regions of code and data.  The code
outside the enclave is disallowed to access the memory inside the
enclave by the CPU access control.

Intel SGX driver provides a ioctl interface for loading and initializing
enclaves and a pager in order to support oversubscription.

This driver is kept under drivers/staging at minimum until root of
trust can be user by using IA32_LEPUBKEYHASHx MSRs.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@linux.intel.com>