kvm/x86: add SGX CPUID support
For KVM_GET_SUPPORTED_CPUID, SGX is reported if hardware supports it. For
KVM_SET_CPUID2, firstly EPC base and size are received via it from Qemu (to
avoid dedicated IOCTL to setup EPC base and size in KVM). Secondly SGX is
disabled if nested is supported. The reason is according to specification,
IA32_VMX_PROCBASED_CTLS2[bit 15] (1-setting of 'enable ENCLS exiting' monitors
CPUID.0x7.0.EBX.SGX, therefore if SGX is exposed via CPUID.0x7.0.EBX.SGX in case
of nested, we need to assume l1 may turn on ENCLS VMEXIT and l0 need to handle
this. But so far as long as oversubscription is not supported, we don't need to
handle ENCLS VMEXIT, for which we disable SGX if nested is supported.

Also add guest_cpuid_has_sgx to indicate whether vcpu supports SGX.

Signed-off-by: Kai Huang <kai.huang@linux.intel.com>